fix 64bit build error

+++ formats/ffmpeg/mm_file_format_ffmpeg_drm.c
@@ -196,7 +196,7 @@ static int mmfile_drm_write(URLContext *h, const unsigned char *buf, int size)
 	return 0;
 }
 
-static long long mmfile_drm_seek(URLContext *h, long long pos, int whence)
+static int64_t mmfile_drm_seek(URLContext *h, long long pos, int whence)
 {
 	MMFileDRMHandle *drmHandle = h->priv_data;
 	drm_trusted_seek_mode_e drm_whence;
